# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InvestRing API",
    "description": "家庭投资组合管理系统 API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/positions": {
      "get": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Get Positions",
        "operationId": "get_positions_api_positions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Portfolio Code"
            }
          },
          {
            "name": "snapshot_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshot Date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Create Position",
        "operationId": "create_position_api_positions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions/{id}": {
      "get": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Get Position",
        "operationId": "get_position_api_positions__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Update Position",
        "operationId": "update_position_api_positions__id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Delete Position",
        "operationId": "delete_position_api_positions__id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions/portfolio/{portfolio_code}/available-cash": {
      "get": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Get Available Cash",
        "operationId": "get_available_cash_api_positions_portfolio__portfolio_code__available_cash_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Code"
            }
          },
          {
            "name": "platform_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions/portfolio/{portfolio_code}/product/{product_code}/available-shares": {
      "get": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Get Available Shares",
        "operationId": "get_available_shares_api_positions_portfolio__portfolio_code__product__product_code__available_shares_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Code"
            }
          },
          {
            "name": "product_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Product Code"
            }
          },
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions/portfolio/{portfolio_code}/investor/{investor_code}/available-shares": {
      "get": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Get Investor Available Shares",
        "operationId": "get_investor_available_shares_api_positions_portfolio__portfolio_code__investor__investor_code__available_shares_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Code"
            }
          },
          {
            "name": "investor_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investor Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions/portfolio/{portfolio_code}/cash-position": {
      "post": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Update Cash Position",
        "description": "更新非净值型资产（现金）金额\n\n权限：仅admin\n规则：\n- 必须在交易日进行\n- 必须指定平台代码\n- 写入 manual_market_value 表（绝对替换），不再直接写 portfolio_position\n- 写入后提示用户重新生成快照（非强制）\n- 同日存在已确认现金交易时附 warnings 提示（issue #88，不阻断）",
        "operationId": "update_cash_position_api_positions_portfolio__portfolio_code__cash_position_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashPositionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "持仓管理"
        ],
        "summary": "List Cash Overrides",
        "description": "查询现金手动覆盖记录（manual_market_value，issue #88）\n\n权限：仅admin",
        "operationId": "list_cash_overrides_api_positions_portfolio__portfolio_code__cash_position_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Code"
            }
          },
          {
            "name": "platform_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform Code"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "持仓管理"
        ],
        "summary": "Delete Cash Override",
        "description": "删除现金手动覆盖记录（issue #88）\n\n删除后该日该平台回退到自然计算值；若覆盖已 baked in 快照，\n需重算快照才生效（requires_snapshot_regen）。\n\n权限：仅admin",
        "operationId": "delete_cash_override_api_positions_portfolio__portfolio_code__cash_position_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Code"
            }
          },
          {
            "name": "platform_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Platform Code"
            }
          },
          {
            "name": "update_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Update Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CashPositionUpdate": {
        "properties": {
          "cash_amount": {
            "type": "number",
            "title": "Cash Amount"
          },
          "platform_code": {
            "type": "string",
            "title": "Platform Code"
          },
          "update_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Update Date"
          }
        },
        "type": "object",
        "required": [
          "cash_amount",
          "platform_code"
        ],
        "title": "CashPositionUpdate",
        "description": "非净值型资产（现金）更新请求"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PositionCreate": {
        "properties": {
          "portfolio_code": {
            "type": "string",
            "title": "Portfolio Code"
          },
          "product_code": {
            "type": "string",
            "title": "Product Code"
          },
          "market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market"
          },
          "platform_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Code"
          },
          "shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares"
          },
          "frozen_shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frozen Shares"
          },
          "cost_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Price"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "market_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value"
          },
          "cash_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Amount"
          },
          "frozen_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frozen Amount"
          },
          "asset_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Type"
          },
          "snapshot_date": {
            "type": "string",
            "format": "date",
            "title": "Snapshot Date"
          }
        },
        "type": "object",
        "required": [
          "portfolio_code",
          "product_code",
          "snapshot_date"
        ],
        "title": "PositionCreate"
      },
      "PositionResponse": {
        "properties": {
          "portfolio_code": {
            "type": "string",
            "title": "Portfolio Code"
          },
          "product_code": {
            "type": "string",
            "title": "Product Code"
          },
          "market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market"
          },
          "platform_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Code"
          },
          "shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares"
          },
          "frozen_shares": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frozen Shares"
          },
          "cost_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Price"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          },
          "market_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Value"
          },
          "cash_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Amount"
          },
          "frozen_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frozen Amount"
          },
          "asset_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Type"
          },
          "snapshot_date": {
            "type": "string",
            "format": "date",
            "title": "Snapshot Date"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name"
          },
          "profit_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profit Loss"
          },
          "profit_loss_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profit Loss Percent"
          }
        },
        "type": "object",
        "required": [
          "portfolio_code",
          "product_code",
          "snapshot_date",
          "id"
        ],
        "title": "PositionResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}